+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginActivity.kt
@@ -49,18 +49,22 @@ class LoginActivity :
     super.onCreate(savedInstanceState)
     loginService.loginActivity = this
     loginViewModel.apply {
-      navigateToHome.observe(
-        this@LoginActivity,
-        {
-          syncBroadcaster.runSync()
-          if (loginViewModel.loginViewConfiguration.value?.enablePin == true) {
-            loginService.navigateToPinLogin(goForSetup = true)
-          } else {
+      navigateToHome.observe(this@LoginActivity) {
+        if (loginViewModel.loginViewConfiguration.value?.enablePin == true) {
+          val lastPinExist = loginViewModel.accountAuthenticator.hasActivePin()
+          if (lastPinExist) {
+            loginViewModel.sharedPreferences.write(FORCE_LOGIN_VIA_USERNAME, false)
+            syncBroadcaster.runSync()
             loginService.navigateToHome()
+          } else {
+            loginService.navigateToPinLogin(goForSetup = true)
           }
+        } else {
+          syncBroadcaster.runSync()
+          loginService.navigateToHome()
         }
-      )
-      launchDialPad.observe(this@LoginActivity, { if (!it.isNullOrEmpty()) launchDialPad(it) })
+      }
+      launchDialPad.observe(this@LoginActivity) { if (!it.isNullOrEmpty()) launchDialPad(it) }
       appLogoResourceFile = getApplicationConfiguration().appLogoIconResourceFile
     }