+++ android/quest/src/test/java/org/smartregister/fhircore/quest/data/PatientRepositoryTest.kt
@@ -32,6 +32,7 @@ import io.mockk.unmockkStatic
 import java.time.LocalDate
 import java.time.ZoneId
 import java.util.Date
+import java.util.Locale
 import javax.inject.Inject
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.runBlockingTest
@@ -40,6 +41,7 @@ import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Condition
 import org.hl7.fhir.r4.model.Encounter
 import org.hl7.fhir.r4.model.Enumerations
+import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.MedicationRequest
 import org.hl7.fhir.r4.model.Observation
 import org.hl7.fhir.r4.model.Patient
@@ -47,6 +49,7 @@ import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.Reference
 import org.hl7.fhir.r4.model.ResourceType
+import org.hl7.fhir.r4.model.StringType
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
@@ -349,7 +352,7 @@ class PatientRepositoryTest : RobolectricTest() {
       )
     }
     with(data.data[0]) {
-      Assert.assertEquals("Questionnaire Name", this[0].value)
+      Assert.assertEquals("Questionnaire Title", this[0].value)
       Assert.assertEquals(" (${today.asDdMmmYyyy()})", this[1].value)
     }
 
@@ -547,5 +550,62 @@ class PatientRepositoryTest : RobolectricTest() {
     coVerify { fhirEngine.search<MedicationRequest>(any()) }
 
     Assert.assertEquals("mr1", result.first().logicalId)
+  }
+
+  @Test
+  fun fetchResultItemLabelShouldReturnLocalizedQuestionnaireTitle() {
+    val questionnaire =
+      Questionnaire().apply {
+        titleElement =
+          StringType("Registration").apply {
+            addExtension(
+              Extension().apply {
+                url = "http://hl7.org/fhir/StructureDefinition/translation"
+                addExtension("lang", StringType("sw"))
+                addExtension("content", StringType("Sajili"))
+              }
+            )
+          }
+
+        nameElement =
+          StringType("Registration2").apply {
+            addExtension(
+              Extension().apply {
+                url = "http://hl7.org/fhir/StructureDefinition/translation"
+                addExtension("lang", StringType("sw"))
+                addExtension("content", StringType("Sajili2"))
+              }
+            )
+          }
+      }
+
+    Locale.setDefault(Locale.forLanguageTag("en"))
+    Assert.assertEquals("Registration", repository.fetchResultItemLabel(questionnaire))
+
+    Locale.setDefault(Locale.forLanguageTag("sw"))
+    Assert.assertEquals("Sajili", repository.fetchResultItemLabel(questionnaire))
+  }
+
+  @Test
+  fun fetchResultItemLabelShouldReturnLocalizedQuestionnaireNameWhenTitleIsAbsent() {
+    val questionnaire =
+      Questionnaire().apply {
+        nameElement =
+          StringType("Registration").apply {
+            addExtension(
+              Extension().apply {
+                url = "http://hl7.org/fhir/StructureDefinition/translation"
+                addExtension("lang", StringType("sw"))
+                addExtension("content", StringType("Sajili"))
+              }
+            )
+          }
+      }
+
+    Locale.setDefault(Locale.forLanguageTag("en"))
+    Assert.assertEquals("Registration", repository.fetchResultItemLabel(questionnaire))
+
+    Locale.setDefault(Locale.forLanguageTag("sw"))
+    Assert.assertEquals("Sajili", repository.fetchResultItemLabel(questionnaire))
   }
 }